// src/app/finance/model/details-ecriture.ts
import { Ecriture } from './ecriture';

export interface DetailsEcriture {
  id?: number;
  typeEcriture?: string;
  detailS1?: string;
  detailS2?: string;
  detailS3?: string;
  detailS4?: string;
  detailS5?: string;
  detailN1?: number;
  detailN2?: number;
  detailN3?: number;
  detailN4?: number;
  detailN5?: number;
  ecriture?: Ecriture;
}